requirements.txt
firebird-driver
pandas
pymssql
pyodbc
